+++ CHANGELOG.md
@@ -1,3 +1,19 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Initial Release
+
+## 0.0.2
+
+* Update README.md
+
+## 0.0.3
+
+* Fix Null Safety
+
+## 0.0.4
+
+* Update README.md and Tutorial
+
+## 0.0.5
+
+* Update Saver to Base64